make initialization logic into provider pattern
