--- example/storybook/stories/Notifications/NotificationScreen.stories.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react-native';
-import { NotificationsScreen } from './';
-
-storiesOf('NotificationsScreen', module).add('demo', () => (
-  <NotificationsScreen />
-));
