+++ example/storybook/stories/Notifications/NotificationsScreen.stories.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react-native';
+import { NotificationsScreen } from '.';
+import { PushNotificationsProvider } from 'src/hooks/usePushNotifications';
+
+storiesOf('NotificationsScreen', module).add('demo', () => (
+  <PushNotificationsProvider>
+    <NotificationsScreen />
+  </PushNotificationsProvider>
+));
