+++ example/storybook/stories/Notifications/index.tsx
@@ -1,21 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import {
-  Text,
-  View,
-  Platform,
-  StyleSheet,
-  TouchableOpacity,
-} from 'react-native';
+import React, { useContext, useEffect } from 'react';
+import { Text, View, StyleSheet, TouchableOpacity } from 'react-native';
 import { Notifications, Notification } from 'react-native-notifications';
 import {
-  getInitialNotification,
-  onNotificationOpened,
-  onNotificationReceived,
   registerDeviceToken,
   requestNotificationsPermissions,
 } from '../../../../src/common/Notifications';
-import { useActiveAccount } from '../../../../src/hooks/useActiveAccount';
-import { useHttpClient } from '../../../../src/hooks/useHttpClient';
+import { PushNotificationsContext } from 'src/hooks/usePushNotifications';
 
 type EventType = 'notificationReceived' | 'notificationOpened';
 
@@ -66,9 +56,13 @@ const styles = StyleSheet.create({
 });
 
 export const NotificationsScreen = () => {
-  const [events, setEvents] = useState<Event[]>([]);
-  const { httpClient } = useHttpClient();
-  const { account } = useActiveAccount();
+  const pushNotificationsContext = useContext(PushNotificationsContext);
+
+  if (!pushNotificationsContext) {
+    return <Text>Context is undefined</Text>;
+  }
+
+  const { events, setEvents, httpClient, account } = pushNotificationsContext;
 
   const sendLocalNotification = () => {
     Notifications.postLocalNotification({
@@ -135,52 +129,6 @@ export const NotificationsScreen = () => {
     });
   }, [account, httpClient]);
 
-  // Set the notification channel for Android
-  useEffect(() => {
-    if (Platform.OS === 'android') {
-      Notifications.setNotificationChannel({
-        channelId: 'sdk-example-channel',
-        name: 'SDK Example',
-        importance: 5,
-        description: 'Channel for the SDK',
-        enableLights: true,
-        enableVibration: true,
-        showBadge: true,
-        vibrationPattern: [200, 1000, 500, 1000, 500],
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    // Handler called when a notification is pressed
-    onNotificationOpened((notification) => {
-      setEvents((events) => [
-        { type: 'notificationOpened', notification },
-        ...events,
-      ]);
-    });
-
-    onNotificationReceived((notification) => {
-      setEvents((events) => [
-        { type: 'notificationReceived', notification },
-        ...events,
-      ]);
-    });
-
-    const getInitial = async () => {
-      // Get the notification that opened the application
-      const notification = await getInitialNotification();
-      if (notification) {
-        setEvents((events) => [
-          { type: 'notificationOpened', notification },
-          ...events,
-        ]);
-      }
-    };
-
-    getInitial();
-  }, []);
-
   return (
     <View>
       <View style={styles.buttonContainer}>
@@ -199,7 +147,10 @@ export const NotificationsScreen = () => {
       </View>
 
       {events.map((event, idx) => (
-        <View key={`event${idx}`}>{renderEvent(event)}</View>
+        <View key={`event${idx}`}>
+          {/* TODO fix typing */}
+          {renderEvent(event as unknown as Event)}
+        </View>
       ))}
     </View>
   );
