+++ example/storybook/stories/index.ts
@@ -2,7 +2,7 @@ import './BrandConfigProvider/BrandConfigProvider.stories';
 import './CustomAppTileScreen.stories';
 import './ExampleApp/ExampleApp.stories';
 import './NoInternetToastProvider.stories';
-import './Notifications/NotificationScreen.stories';
+import './Notifications/NotificationsScreen.stories';
 import './OAuth.stories';
 import './ActivityIndicatorView.stories';
 import './TrackTile/AdvancedTrackerDetails.stories';
